Rename RecursiveObservable to DeepObservable
And also make this depend on the newer ObservableObject and
ObservableArray. Closes #3 and #4.
