+++ CONTRIBUTING.md
@@ -1,8 +1,8 @@
-# Contributing to can-recursive-observable
+# Contributing to can-deep-observable
 
 ## Developing Locally
 
-This section will walk you through setting up the [repository](https://github.com//can-recursive-observable) on your computer.
+This section will walk you through setting up the [repository](https://github.com/canjs/can-deep-observable) on your computer.
 
 ### Signing up for GitHub
 
@@ -12,14 +12,14 @@ If you don’t already have a GitHub account, you’ll need to [create a new one
 
 A “fork” is a copy of a repository in your personal GitHub account. “Cloning” is the process of getting the repository’s source code on your computer.
 
-GitHub has a guide for [forking a repo](https://help.github.com/articles/fork-a-repo/). To fork can-recursive-observable, you can start by going to its [fork page](https://github.com//can-recursive-observable/fork).
+GitHub has a guide for [forking a repo](https://help.github.com/articles/fork-a-repo/). To fork can-deep-observable, you can start by going to its [fork page](https://github.com/canjs/can-deep-observable/fork).
 
 Next, you’ll want to clone the repo. [GitHub’s cloning guide](https://help.github.com/articles/cloning-a-repository/) explains how to do this on Linux, Mac, or Windows.
 
 GitHub’s guide will [instruct you](https://help.github.com/articles/fork-a-repo/#step-2-create-a-local-clone-of-your-fork) to clone it with a command like:
 
 ```shell
-git clone https://github.com/YOUR-USERNAME/can-recursive-observable
+git clone https://github.com/YOUR-USERNAME/can-deep-observable
 ```
 
 Make sure you replace `YOUR-USERNAME` with your GitHub username.
@@ -33,7 +33,7 @@ First, make sure you’ve [installed Node.js and npm](https://docs.npmjs.com/get
 If you just cloned the repo from the command line, you’ll want to switch to the folder with your clone:
 
 ```shell
-cd can-recursive-observable
+cd can-deep-observable
 ```
 
 Next, install the project’s dependencies with npm:
