+++ can-deep-observable.md
@@ -1,21 +1,21 @@
-@module {function} can-recursive-observable
+@module {function} can-deep-observable
 @parent can-observables
 @collection can-ecosystem
-@alias can.RecursiveObservable
+@alias can.DeepObservable
 @package ./package.json
 @outline 2
 @templateRender true
 
 @description Create observable objects where nested objects and arrays are also observable.
 
-@signature `new RecursiveObservable(data)`
+@signature `new DeepObservable(data)`
 
 Creates an observable object with `data`, where inner objects and arrays are converted to [can-observable-object] and [can-observable-array] types.
 
 ```js
-import { RecursiveObservable } from "can";
+import { DeepObservable } from "can";
 
-let obj = new RecursiveObservable({
+let obj = new DeepObservable({
   inner: {
     prop: "value"
   }
@@ -30,18 +30,18 @@ obj.inner.prop = "new value";
 @codepen
 
   @param {Object} A plain object from which observable values will derive.
-  @return {can-recursive-observable} A new `RecursiveObservable` object.
+  @return {can-deep-observable} A new `DeepObservable` object.
 
-@signature `prop: RecursiveObservable`
+@signature `prop: DeepObservable`
 
-Create a new property on [can-observable-object] that is a recursive object.
+Create a new property on [can-observable-object] that is a deep observable object.
 
 ```js
-import { ObservableObject, RecursiveObservable } from "can";
+import { ObservableObject, DeepObservable } from "can";
 
 class Faves extends ObservableObject {
   static props = {
-    config: RecursiveObservable
+    config: DeepObservable
   };
 }
 
@@ -65,16 +65,16 @@ faves.config.inner.config = { enabled: true };
 
 ## Use
 
-__RecursiveObservable__ is a type that turns an object, and nested objects, into [can-observable-object ObservableObjects] and [can-observable-array ObservableArrays].
+__DeepObservable__ is a type that turns an object, and nested objects, into [can-observable-object ObservableObjects] and [can-observable-array ObservableArrays].
 
-It's most useful as a way to take plain objects which have no definable structure, but nevertheless need to be observable. Since you can't predefine these objects on [can-observable-object ObservableObjects], you can instead use RecursiveObject as the type.
+It's most useful as a way to take plain objects which have no definable structure, but nevertheless need to be observable. Since you can't predefine these objects on [can-observable-object ObservableObjects], you can instead use DeepObservable as the type.
 
 ```js
-import { ObservableObject, RecursiveObservable, restModel } from "can";
+import { ObservableObject, DeepObservable, restModel } from "can";
 
 class ServiceData extends ObservableObject {
   static props = {
-    options: RecursiveObservable
+    options: DeepObservable
   }
 }
 
@@ -86,5 +86,4 @@ ServiceData.connection = restModel({
 ServiceData.get({ id: "1" }).then(data => {
   data.options.nested.props...
 });
-
 ```
