+++ test.html
@@ -1,4 +1,4 @@
 <!doctype html>
-<title>can-recursive-observable</title>
-<script src="node_modules/steal/steal.js" main="~/can-recursive-observable-test"></script>
+<title>can-deep-observable tests</title>
+<script src="node_modules/steal/steal.js" main="~/can-deep-observable-test"></script>
 <div id="qunit-fixture"></div>
